refactor(ndk): separate event reading funcs into own file
