+++ bugsnag-plugin-android-ndk/src/main/jni/utils/serializer/migrate.h
